requirements.txt
attrs==20.2.0
awsebcli==3.19.2
bcrypt==3.2.0
blessed==1.17.11
botocore==1.19.6
cached-property==1.5.2
cement==2.8.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==3.2
docker==4.3.1
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
Flask==1.1.2
future==0.16.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
MarkupSafe==1.1.1
paramiko==2.7.2
pathspec==0.5.9
psycopg2-binary==2.8.6
pycparser==2.20
PyNaCl==1.4.0
pyrsistent==0.17.3
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.24.0
semantic-version==2.5.0
six==1.15.0
termcolor==1.1.0
texttable==1.6.3
urllib3==1.25.11
wcwidth==0.1.9
websocket-client==0.57.0
Werkzeug==1.0.1
